% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\alias{chart.default}
\alias{chart.function}
\alias{$.subsettable_type}
\alias{chart}
\alias{combine_charts}
\title{Create charts}
\usage{
chart(data, ..., type = NULL, env = parent.frame())

\method{chart}{default}(data, specif = NULL, formula = NULL,
  mapping = NULL, ..., type = NULL, auto.labs = TRUE,
  env = parent.frame())

\method{chart}{function}(data, ..., type = NULL, auto.labs = TRUE,
  env = parent.frame())

\method{$}{subsettable_type}(x, name)

combine_charts(chartlist, ncol = NULL, nrow = NULL, labels = "AUTO", ...)
}
\arguments{
\item{data}{The dataset (a \code{data.frame} or \code{tibble}, usually).}

\item{...}{Further arguments.}

\item{type}{The type of plot to produce.}

\item{env}{The environment where to evaluated the formula.}

\item{specif}{Specification, being either \code{aes()}, or a formula.}

\item{formula}{A formula.}

\item{mapping}{An \code{aes()} object, as for \code{\link[=ggplot]{ggplot()}}.}

\item{auto.labs}{Are labels (and units) automatically used for axes?}

\item{x}{A \code{subsettable_type} function.}

\item{name}{The value to use for the \code{type=} argument.}

\item{chartlist}{List of charts to combine.}

\item{ncol}{(optional) number of columns in the plot grid.}

\item{nrow}{(optional) number of rows in the plot grid.}

\item{labels}{(optional) labels to use for each individual plot. \code{"AUTO"}
(default value) auto-generates uppercase labels, and \code{"auto"} does the same
for lowercase labels.}
}
\value{
An object of class \code{ggarrange} containing a list of \code{ggplot}s.
}
\description{
\code{chart()} provides a unified interface for base plots, lattice and ggplot2.

Assemble multiple charts on the same page. Wrapper arround \code{\link[=ggarrange]{ggarrange()}} with
different defaults.
}
\details{
....
}
\examples{
urchin <- data::read("urchin_bio", package = "data", lang = "en")

# base R graphics
hist(urchin$height)
# ... translates to:
chart(function() hist(urchin$height))
# ... or if the expression is provided directly, specify it is a base chart
chart(hist(urchin$height), type = "base")
# ... or more concisely:
chart$base(hist(urchin$height))

# A lattice plot
histogram(~ height, data = urchin)
chart$histogram(urchin, ~ height)

# ggplot2 histogram
ggplot(urchin, aes(height)) + geom_histogram()
#... or with chart (notice similarities with lattice version)
chart(urchin,  ~ height) + geom_histogram()
chart$geom_histogram(urchin,  ~ height)
urchin <- data::read("urchin_bio", package = "data", lang = "en")
}
\seealso{
\code{\link[=f_aes]{f_aes()}}, \code{\link[=ggplot]{ggplot()}}

\code{\link[=chart]{chart()}}, \code{\link[=ggarrange]{ggarrange()}}
}
\concept{
R plots with graphics, lattice or ggplot2

Combine plots on the same page
}
\keyword{hplot}
