% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_theme.R
\docType{data}
\name{chart_theme}
\alias{chart_theme}
\alias{theme_sciviews}
\alias{theme_sciviews_lattice}
\alias{theme_sciviews_graphics}
\alias{theme_svgray}
\alias{theme_svgray_lattice}
\alias{theme_svgray_graphics}
\alias{theme_svmap}
\alias{theme_svmap_lattice}
\alias{theme_svmap_graphics}
\title{Select chart theme}
\format{An object of class \code{NULL} of length 0.}
\usage{
chart_theme(theme)

theme_sciviews(font_size = 12, font_family = "", line_size = 0.5)

theme_sciviews_lattice(...)

theme_sciviews_graphics

theme_svgray()

theme_svgray_lattice(...)

theme_svgray_graphics()

theme_svmap()

theme_svmap_lattice()

theme_svmap_graphics()
}
\arguments{
\item{theme}{The theme to apply (character string).}

\item{font_size}{The default size font for this theme.}

\item{font_family}{The default font family in this theme.}

\item{line_size}{The default line size in this theme.}

\item{...}{Arguments passed to \code{\link[=ggplot2like]{ggplot2like()}}, the most used being \code{n=} for
the number of colours to generate in the palette.}
}
\description{
Lattice and ggplot2 provide themes for their plots, while with base R plots,
one specifies appearance through \code{\link[=par]{par()}}. \code{chart_theme()} tries to get the
plot appearance as uniform as possible beteen the three plot engines. So,
setting themes this way change the appearance of all three kinds of plots.
}
\examples{
# TODO..
}
\seealso{
\code{\link[=chart]{chart()}}, \code{\link[=theme]{theme()}}, \code{\link[=trellis.par.set]{trellis.par.set()}}
}
\concept{
Themes for R plots with graphics, lattice or ggplot2
}
\keyword{color}
\keyword{datasets}
