% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_aes.R
\name{f_aes}
\alias{f_aes}
\title{Create aes()thetics from formula for ggplot2}
\usage{
f_aes(formula, ..., with.facets = FALSE)
}
\arguments{
\item{formula}{A formula.}

\item{...}{Further aesthetics to set (like \code{size}, \code{colour}, et.)}

\item{with.facets}{Do we create special (non-ggplot2) aesthetics for facets
(no by default)?}
}
\value{
An aesthetic object of class \code{uneval}, as those obtained with \code{\link[=aes]{aes()}}.
}
\description{
This function allows to use a formula interface directly with \strong{ggplot2}, or
\strong{chart}.
}
\examples{
# TODO...
}
