% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sgg.R
\name{Sgg}
\alias{Sgg}
\alias{$.subsettable_Sgg}
\alias{.DollarNames.subsettable_Sgg}
\title{Make a ggplot function pipeable}
\usage{
Sgg(ggplot, ...)

\method{$}{subsettable_Sgg}(x, name)

\method{.DollarNames}{subsettable_Sgg}(x, pattern = "")
}
\arguments{
\item{ggplot}{An object of class "ggplot" (or "theme").}

\item{...}{Further arguments passed to the the ggplot function (see Details).}

\item{x}{The \code{Sgg()}function.}

\item{name}{The name of the ggplot function to make pipeable.}

\item{pattern}{A regular expression to list matching names.}
}
\value{
The \code{Sgg()} function just returns an error message. When subsetted
with the name of a {ggplot2} function (e.g., \code{Sgg$geom_point()}), it
returns a modified version of that function in such a way that it can be
used with a pipe operator.
}
\description{
The set \code{Sgg} should be used like this: \code{Sgg$geom_point()}.
It transforms on the fly an original {ggplot2} function supposed to be used
with the \code{+} operator (like \code{p + geom_point()}) into a pipeable version
(like \code{p \%>\% Sgg$geom_point()}).
}
\details{
The function returned by \code{Sgg$fun} is a modified version of the
function \code{fun} where a first argument \verb{ggplot =} is added, and the
instruction \code{ggplot + ...} is added in its body. A message
is also added in the body to explicitly warn about these changes. All the
other arguments of \code{fun} remain valid and should keep their original meaning.

The changes are done on the fly, and the original function \code{fun} is \strong{not}
altered anywhere else (and in particular, no alteration is done in a package
or a namespace). When using this construct, make sure that: (1) you
understand what is done, (2) you are aware that you use an altered version of
the original function, (3) a bug or strange behavior may occur due to the
patch and the original author of the function is not responsible in this case
(the problem must be reported to the author of \code{Sgg} and the maintainer of
the present package instead), and (4) the patched function exhibits an
additional argument and behaves differently to what is described in the help
page of the original, non-patched, function!
}
\examples{
library(ggplot2)
data(iris)
ggplot(aes(x = Petal.Length, y = Sepal.Length, col = Species) , data = iris) |>
  Sgg$geom_point() |>
  Sgg$labs(x = "Sepal length (mm)", y = "Petal length (mm)")
# Also try completion with Sgg$<tab>
}
