% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-package.R
\docType{package}
\name{chart-package}
\alias{chart-package}
\title{Chart, unified interface for R plots}
\description{
Unification of base plots, lattice and ggplot2, providing a formula interface
for all three plot engines.
}
\section{Important functions}{

\itemize{
\item \code{\link[=chart]{chart()}} constructs a \code{Chart} object.
}
}

