% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\alias{chart.default}
\alias{chart.function}
\alias{$.subsettable_type}
\title{Create charts}
\usage{
chart(data, ..., type = NULL, env = parent.frame())

\method{chart}{default}(data, specif = NULL, formula = NULL,
  mapping = NULL, ..., type = NULL, auto.labs = TRUE,
  env = parent.frame())

\method{chart}{function}(data, ..., type = NULL, auto.labs = TRUE,
  env = parent.frame())

\method{$}{subsettable_type}(x, name)
}
\arguments{
\item{data}{The dataset (a \code{data.frame} or \code{tibble}, usually).}

\item{...}{Further arguments.}

\item{type}{The type of plot to produce.}

\item{env}{The environment where to evaluated the formula.}

\item{specif}{Specification, being either \code{aes()}, or a formula.}

\item{formula}{A formula.}

\item{mapping}{An \code{aes()} object, as for \code{\link[=ggplot]{ggplot()}}.}

\item{auto.labs}{Are labels (and units) automatically used for axes?}

\item{x}{A \code{subsettable_type} function.}

\item{name}{The value to use for the \code{type=} argument.}
}
\description{
\code{chart()} provides a unified interface for base plots, lattice and ggplot2.
}
\details{
....
}
\examples{
# TODO..

# base R graphics
chart(function() hist(rnorm(50)))
# ... or if the expression is provided directly, specify it is a base chart
chart(hist(rnorm(50)), type = "base")
# ... or more concisely:
chart$base(hist(rnorm(50)))

# A lattice plot
chart$histogram(iris, ~ Sepal.Length)
}
\seealso{
\code{\link[=f_aes]{f_aes()}}, \code{\link[=ggplot]{ggplot()}}
}
\concept{
R plots with graphics, lattice or ggplot2
}
\keyword{hplot}
