% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\alias{chart.default}
\title{Create charts}
\usage{
chart(data, ..., type = NULL, env = parent.frame())

\method{chart}{default}(data, specif = NULL, formula = NULL,
  mapping = NULL, ..., type = NULL, env = parent.frame())
}
\arguments{
\item{data}{The dataset (a \code{data.frame} or \code{tibble}, usually).}

\item{...}{Further arguments.}

\item{type}{The type of plot to produce.}

\item{env}{The environment where to evaluated the formula.}

\item{specif}{Specification, being either \code{aes()}, or a formula.}

\item{formula}{A formula.}

\item{mapping}{An \code{aes()} object, as for \code{\link[=ggplot]{ggplot()}}.}
}
\description{
\code{chart()} provides a unified interface for base plots, lattice and ggplot2.
}
\details{
....
}
\examples{
# TODO..
}
\seealso{
\code{\link[=f_aes]{f_aes()}}, \code{\link[=ggplot]{ggplot()}}
}
\concept{
R plots with graphics, lattice or ggplot2
}
\keyword{hplot}
